UNIV4 SWAP/requirements.txt
python-telegram-bot>=20.0
web3>=6.0.0
python-dotenv>=1.0.0
eth-abi>=4.0.0 